Humanoid-Robotics/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client>=1.7.0
openai==0.28.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
markdown==3.5.1
beautifulsoup4==4.12.2
tiktoken==0.5.2
asyncio==3.4.3
httpx==0.25.2